requirements.txt
Flask
PyPDF2
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
gunicorn
Werkzeug
